v1.0.1
- Switched program icon to an Alien
- Moved global.hpp out of subsystemHeaders
- Slight change to input curving for increased precision
- Fixed motor temperature check
- Expansion can now only activate when 15 seconds are left in the Driver Control Period
- Controller rumbles when activating expansion

+++ src/subsystemFiles/drive.cpp
@@ -1,7 +1,7 @@
 #include "main.h"
 
 float getMotorOutput(int i) {
-    return i*exp((float)((abs(i)-100)*60)/1000);
+    return i*exp((double)((abs(i)-100)*60)/1000);
 }
 
 void DirectionToggle() {
